pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Multi-Modal Local RAG with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.13.2",
    "hf-xet>=1.2.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "pdf2image>=1.17.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "unstructured[all-docs]>=0.18.21",
]
[project.optional-dependencies]
vector-faiss = ["faiss-cpu"]
